n64: fix analog stick inputs on arm64
Converting a negative floating-point value to an unsigned integer is
undefined behavior and just happens to do the desired thing when
compiling for x86-64. Convert to a signed integer first to avoid UB and
guarantee the expected result on arm64.

+++ ares/n64/controller/gamepad/gamepad.cpp
@@ -229,8 +229,8 @@ auto Gamepad::read() -> n32 {
   }
 
   n32 data;
-  data.byte(0) = -ay;
-  data.byte(1) = +ax;
+  data.byte(0) = s8(-ay);
+  data.byte(1) = s8(+ax);
   data.bit(16) = cameraRight->value();
   data.bit(17) = cameraLeft->value();
   data.bit(18) = cameraDown->value();
